developing EDA funcs
